usb:Ps5169 probe issue
The chip id shall be two bytes
Otherwise their may be some garbage data
In the memory and cause probe failure.

Reviewed-on: https://gerrit.mot.com/1989133
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Submit-Approved: Jira Key

+++ drivers/usb/misc/ssusb-redriver-ps5169.c
@@ -966,7 +966,7 @@ static int ps5169_i2c_probe(struct i2c_client *client,
 	struct device *dev = &client->dev;
 	struct ps5169_redriver *ps5169;
 	int ret, size = 0;
-	u32 device_id;
+	u16 device_id;
